requirements.txt
transformers
pillow
opencv-python
torch
